src/app/requirements.txt
dash
plotly 
pandas
gunicorn==20.0.4
psycopg2-binary